drive contents from db being rendered
