+++ src/app/file-row.tsx
@@ -1,8 +1,11 @@
-import type { Folder, File } from "../lib/mock-data";
 import { Folder as FolderIcon, FileIcon } from "lucide-react";
 import Link from "next/link";
+import type { files, folders } from "~/server/db/schema";
 
-export function FileRow(props: { file: File; handleFileClick: () => void }) {
+export function FileRow(props: {
+  file: typeof files.$inferSelect;
+  handleFileClick: () => void;
+}) {
   const { file, handleFileClick } = props;
   return (
     <li
@@ -28,7 +31,7 @@ export function FileRow(props: { file: File; handleFileClick: () => void }) {
 }
 
 export function FolderRow(props: {
-  folder: Folder;
+  folder: typeof folders.$inferSelect;
   handleFolderClick: () => void;
 }) {
   const { folder, handleFolderClick } = props;
