+++ src/server/db/schema.ts
@@ -1,3 +1,5 @@
+import "server-only";
+
 import {
   bigint,
   int,
